pyproject.toml
[tool.poetry]
name = "empi"
version = "0.1.0"
description = "Backend for the Empirical research CMS"
authors = ["David Krchňavý <krchnavy11@uniba.sk>"]
readme = "README.md"

[tool.poetry.urls]
"Homepage" = "https://gitlab.matfyz.sk/znamafirma/empi"
"Bug Tracker" = "https://gitlab.matfyz.sk/znamafirma/empi/-/issues"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.15.1"
pycryptodome = "^3.20.0"
python-magic = "^0.4.27"
django-rest-knox = "^4.2.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}
gunicorn = "^22.0.0"
nanoid = "^2.0.0"
html2text = "^2024.2.26"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
